+++ src/main/java/bridge/Constants.java
@@ -2,9 +2,11 @@ package bridge;
 
 public class Constants {
 
-  public static final String RESTART = "R";
-  public static final String STOP = "Q";
-  public static final String SUCCESS = "성공";
-  public static final String FAIL = "실패";
+  public static final String BRIDGE_UPPER = "U";
+  public static final String BRIDGE_LOWER = "D";
+  public static final String GAME_RESTART = "R";
+  public static final String GAME_QUIT = "Q";
+  public static final String GAME_SUCCESS = "성공";
+  public static final String GAME_FAIL = "실패";
 
 }
