+++ src/main/java/bridge/vo/Cross.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.vo;
 
 public class Cross {
