+++ src/main/java/bridge/vo/Direction.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.vo;
 
 import java.util.Arrays;
